Split plugin config facade
